refactored state logic into Home Page. Removed the need for container components for List and AddItem components.
