+++ frontend/src/components/Items/List/ItemList.tsx
@@ -1,31 +1,75 @@
 import * as React from 'react';
 
-import { useQuery, useMutation } from '@apollo/react-hooks';
+import { ApolloError } from 'apollo-boost';
 
-import ItemListView from './ItemListView';
+import { makeStyles } from '@material-ui/core/styles';
+import Typography from '@material-ui/core/Typography';
+import IconButton from '@material-ui/core/IconButton';
+import DeleteIcon from '@material-ui/icons/Delete';
+import Divider from '@material-ui/core/Divider';
 
-import { ITEMS_QUERY, DELETE_ITEM_MUTATION } from '../queries';
+import { IItem } from '../../../types/item';
 
 interface IPropsType {
-  searchCategory?: string;
+  loading: boolean;
+  error: ApolloError | undefined;
+  items: IItem[];
+  onDeleteItem: (id: string) => void;
 }
 
-const ItemList = ({ searchCategory }: IPropsType) => {
-  const { loading, error, data } = useQuery(ITEMS_QUERY, {
-    variables: { category: searchCategory }
-  });
-  const [deleteItem] = useMutation(DELETE_ITEM_MUTATION, {
-    refetchQueries: [{ query: ITEMS_QUERY }]
-  });
+const useStyles = makeStyles({
+  container: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    margin: '1rem 0',
+    width: '100%'
+  },
+  row: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-around',
+    width: '50%'
+  },
+  col: {
+    width: '25%'
+  }
+});
 
-  const onDeleteItem = async (_id: string) => {
-    deleteItem({ variables: { _id } });
-  };
+const ItemListView = ({ loading, error, items, onDeleteItem }: IPropsType) => {
+  const classes = useStyles();
 
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error :(</p>;
+  if (loading && !items) return <p>Loading items...</p>;
+  if (error) return <p>An error occured fetching items.</p>;
 
-  return <ItemListView items={data.getItems} onDeleteItem={onDeleteItem} />;
+  return (
+    <div className={classes.container}>
+      <div className={classes.row}>
+        <Typography className={classes.col} variant="h6">
+          Name
+        </Typography>
+        <Typography className={classes.col} variant="h6">
+          Category
+        </Typography>
+        <Typography className={classes.col} variant="h6">
+          Price
+        </Typography>
+        <Typography variant="h6">Actions</Typography>
+      </div>
+      <Divider color="primary" />
+      {items.map((item: IItem) => (
+        <div key={item._id} className={classes.row}>
+          <Typography className={classes.col}>{item.name}</Typography>
+          <Typography className={classes.col}>{item.category}</Typography>
+          <Typography className={classes.col}>{item.price}</Typography>
+          <IconButton color="primary" onClick={() => onDeleteItem(item._id)}>
+            <DeleteIcon />
+          </IconButton>
+        </div>
+      ))}
+    </div>
+  );
 };
 
-export default ItemList;
+export default ItemListView;
